Create user_guide.md
This commit introduces a detailed User Guide in Markdown format for the Surjo Chrome Extension. 

The purpose of this User Guide is to assist users in understanding how to install, configure, and use the Surjo Chrome Extension effectively. It provides step-by-step instructions and helpful tips to ensure that users can leverage all the features of the extension to the fullest extent.